Update business hour check logic test cases to accomodate holiday list
